[hotfix] Factor out AbstractImitatingJobManagerMetricGroup super class
The AbstractImitatingJobManagerMetricGroup imitates the reporting of the
JobManagerMetricGroup and can be used for metrics which are factored out
of the JobManagerMetricGroup without breaking backwards compatibility wrt
the reported metrics.
